+++ launching/src/main/java/org/jboss/tools/ssp/eclipse/jdt/internal/launching/StandardVMRunner.java
@@ -38,9 +38,9 @@ import org.jboss.tools.ssp.eclipse.jdt.launching.IVMInstall;
 import org.jboss.tools.ssp.eclipse.jdt.launching.StandardVMType;
 import org.jboss.tools.ssp.eclipse.jdt.launching.VMRunnerConfiguration;
 import org.jboss.tools.ssp.eclipse.osgi.util.NLS;
+import org.jboss.tools.ssp.internal.launching.LaunchingCore;
 import org.jboss.tools.ssp.internal.launching.util.NativeEnvironmentUtils;
 import org.jboss.tools.ssp.internal.launching.util.OSUtils;
-import org.jboss.tools.ssp.launching.LaunchingCore;
 
 /**
  * A launcher for running Java main classes.
