+++ launching/src/main/java/org/jboss/tools/ssp/internal/launching/RunningVMSyspropCache.java
@@ -6,7 +6,7 @@
  * 
  * Contributors: Red Hat, Inc.
  ******************************************************************************/
-package org.jboss.tools.ssp.launching;
+package org.jboss.tools.ssp.internal.launching;
 
 import java.util.HashMap;
 import java.util.Map;
