+++ server-wildfly/src/main/java/org/jboss/tools/ssp/server/wildfly/servertype/impl/JBossServerDelegate.java
@@ -22,7 +22,7 @@ import org.jboss.tools.ssp.eclipse.debug.core.ILaunch;
 import org.jboss.tools.ssp.eclipse.debug.core.model.IProcess;
 import org.jboss.tools.ssp.eclipse.jdt.launching.IVMInstall;
 import org.jboss.tools.ssp.eclipse.jdt.launching.IVMInstallRegistry;
-import org.jboss.tools.ssp.launching.LaunchingCore;
+import org.jboss.tools.ssp.internal.launching.LaunchingCore;
 import org.jboss.tools.ssp.server.ServerCoreActivator;
 import org.jboss.tools.ssp.server.model.AbstractServerDelegate;
 import org.jboss.tools.ssp.server.spi.model.polling.IPollResultListener;
